refactor(test): extract test related to generics out of internal/external module tests
