+++ java2typescript-jackson/src/test/resources/java2typescript/jackson/module/ClassWithGenericTypeTest.classExtendsClassWithGenericTypeParams.d.ts
@@ -1,0 +1,4 @@
+export interface StringClass {
+    someField: string;
+    genericList: string[];
+}
