+++ src/test/resources/java2typescript/jackson/module/DefinitionGeneratorTest.externalModuleFormat.d.ts
@@ -1,7 +1,3 @@
-export interface StringClass {
-    someField: string;
-}
-
 export enum ChangedEnumName {
     VAL1,
     VAL2,
